Working on text rendering overlays
