+++ interface/src/avatar/Avatar.cpp
@@ -281,7 +281,7 @@ enum TextRendererType {
 
 static TextRenderer3D* textRenderer(TextRendererType type) {
     static TextRenderer3D* chatRenderer = TextRenderer3D::getInstance(SANS_FONT_FAMILY, -1,
-        false, TextRenderer3D::SHADOW_EFFECT);
+        false, SHADOW_EFFECT);
     static TextRenderer3D* displayNameRenderer = TextRenderer3D::getInstance(SANS_FONT_FAMILY);
 
     switch(type) {
