--- libraries/render-utils/src/sdf_text.slv
@@ -1,24 +1,0 @@
-<@include gpu/Config.slh@>
-<$VERSION_HEADER$>
-//  Generated on <$_SCRIBE_DATE$>
-//  sdf_text.vert
-//  vertex shader
-//
-//  Created by Brad Davis on 10/14/13.
-//
-//  Distributed under the Apache License, Version 2.0.
-//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
-//
-
-uniform mat4 Projection;
-uniform mat4 ModelView;
-
-attribute vec2 Position;
-attribute vec2 TexCoord;
-
-varying vec2 vTexCoord;
-
-void main() {
-  vTexCoord = TexCoord;
-  gl_Position = Projection * ModelView * vec4(Position, 0.0, 1.0);
-}
